created atom for dialog and abstracted logic to hook
