+++ atoms/showUpdateDialogAtom.tsx
@@ -1,0 +1,6 @@
+import { atom } from 'recoil'
+
+export const showUpdateDialogState = atom({
+  key: 'showUpdateDialogStateshowUpdateDialogState',
+  default: false
+})
