+++ components/DialogUpdateProduct.tsx
@@ -1,35 +1,20 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { Dialog } from 'primereact/dialog'
 import { InputText } from 'primereact/inputtext'
-import React, { useState } from 'react'
-import { useRecoilState } from 'recoil'
-import { Product } from '../@types/frontend.types'
-import { selectedProductState } from '../atoms/selectedProductAtom'
-import { updateProduct } from '../services/updateProduct'
+import React from 'react'
 
-const DialogUpdateProduct = ({ showUpdateDialog, setShowUpdateDialog }) => {
-  const [selectedProduct, setSelectedProduct] =
-    useRecoilState(selectedProductState)
-  const [productName, setProductName] = useState('')
-  const [productPrice, setProductPrice] = useState(0)
-  const queryClient = useQueryClient()
-  const updateQuery = ({ id, name, price }: Product) =>
-    updateProduct({ id, name, price })
-  const { mutate } = useMutation(updateQuery, {
-    onSuccess: () => {
-      // Invalidate and refetch
-      queryClient.invalidateQueries(['products'])
-      setShowUpdateDialog(false)
-      setSelectedProduct({
-        id: '',
-        name: '',
-        price: 0
-      })
-    }
-  })
-  const handleUpdateProduct = () => {
-    mutate({ id: selectedProduct.id, name: productName, price: productPrice })
-  }
+import useDialogUpdateProduct from '../hooks/useDialogUpdateProduct'
+
+const DialogUpdateProduct = () => {
+  const {
+    handleUpdateProduct,
+    productName,
+    productPrice,
+    setProductName,
+    setProductPrice,
+    showUpdateDialog,
+    setShowUpdateDialog
+  } = useDialogUpdateProduct()
+
   return (
     <Dialog
       visible={showUpdateDialog}
